update all db but in ignore

+++ update_schema.php
@@ -7,6 +7,10 @@ printVariablesFromFile('parameters.php');
 
 $conn = connectToServer($servername, $username, $password);
 
+if ($listDatabases)  {
+    listDatabases($databasePrefix, $conn);
+ }
+
 // *** START initial checks ***
 
 // target databases
@@ -21,13 +25,14 @@ if (!$UPDATE_ALL_DATABASES) {
     } else {
         echo "All target databases are present\n";
     }
+    echo "\n";
 }
 
 // ignore databases
 if ($UPDATE_ALL_DATABASES) {
 // only if we update ALL the databases we check about missing ignore databases
 
-    $db = databaseNamePrefixId($ignoreDatabase);
+    $db = databaseNamePrefixId($ignoreDatabase, $databasePrefix);
     $missingDatabases = databaseCheckMissing($conn, $db);
 
     if (!empty($missingDatabases)) {
@@ -35,9 +40,14 @@ if ($UPDATE_ALL_DATABASES) {
     } else {
         echo "All ignore databases are present\n";
     }
+    echo "\n";
 }
 
 //alter  query
+if (!isset($alterQuery)) {
+    die("Missing ALTER query");
+}
+
 if(!$ALLOW_DROP_COLUMN ) {
 //if we are not allowed to drop column, check whether it is present
     if (stringContains($alterQuery, 'DROP COLUMN')) {
@@ -51,17 +61,21 @@ if(!$ALLOW_DROP_TABLE ) {
             die("Settings do not allow DROP TABLE");
         }
 }
-
 // *** END initial checks ***
 
 if (!$UPDATE_ALL_DATABASES && !empty($targetDatabase)) {
+// we update target tables
     $db = databaseNamePrefixId($targetDatabase);
     updateDatabases($db, $alterQuery, $conn, $servername,  $username, $password, $test);
 }
 
-
- if ($listDatabases)  {
-    listDatabases($databasePrefix, $conn);
+if ($UPDATE_ALL_DATABASES && !empty($ignoreDatabase)) {
+// we update ALL tables except ignore tables
+        $matchingDatabases = getDatabasesWithPrefix($conn);
+        $ignoreDatabases = databaseNamePrefixId($ignoreDatabase);
+        $notInIgnore = array_diff($matchingDatabases, $ignoreDatabases);
+        //echo  implode(", ", $notInIgnore);
+        updateDatabases($notInIgnore, $alterQuery, $conn, $servername,  $username, $password, $test);
  }
 
 $conn->close();
